Updated Lesson Page, Addition to About-Us Page
Added some navigation functionality to the first page of the lesson page. Will still need a bit of deliberation, but it's taking shape.

Tidied up subject.html and added a link to an upcoming lesson page.

Populated the timeline on the about-us page. More to come, but the idea is there. Uses Google Spreadsheets to update so it's easy for everyone.

Signed-off-by: Kenneth Lim <kenlimmj@gmail.com>