Finalized Contact Us, Team, Partners Page.
NO MORE CHANGES. NO MORE. OH GOD NO MORE CHANGES.

Standardized the design across the pages and added a universal navigation on the top right. Press kit is skimpy and only has the timeline, but WE DON'T NEED NO FUCKING PRESS KIT. (yet)

So we'll lock this up and consider it final. So there.

Signed-off-by: Kenneth Lim <kenlimmj@gmail.com>