Massive Updates
Added a drag-and-drop interface to manage-economics.html and manage-lesson.html. Uses the native HTML5 DnD API. Good practice suggests using Modernizr to check if the browser supports the DnD, else to polyfill, but fuck that shit. If someone working in OL uses IE or an outdated version of FF/Chrome/Opera, they are fired.

Added checkpoint.html, topic.html and lesson.html. These are add/edit pages for their respective titles. Has a preview section which should be live-updated using whatever form of on-the-wire magic we are using.

Updated login.html with HTML5 form goodness.

Added a draft table listing to the staff.html page. Still missing a search bar (to be implemented with list.js) and the rest of the people. Will also create user.html to add user.

Signed-off-by: Kenneth Lim <kenlimmj@gmail.com>