I DON'T WANNA DO THIS ANYMORE.
Standardized the navbar on the left across all the pages.

Starting page for checkpoints in reveal.js slide will probably also be standardized this way.

Added a sticky to the navbar so it stays in the viewport on scroll.

Still a lot of inconsistencies in the code, especially in header and footer. Pressing ahead first and will return to touch up when shipping for production.

FML I am going to hell for my shit code. Inline styles everywhere, zero abstraction. Zero library.

Done up opportunities.html, partners.html and about-us.html.

Left with contact-us.html and press.html. After that, and the lesson page. WE ARE DONE.

Oh wait. Home page. Yeah. And that.

Signed-off-by: Kenneth Lim <kenlimmj@gmail.com>