Lesson Page Update + Final Commit
Updated the front of the lesson page with the necessary items for design consistency.

Everything is now final and good for integration. The front page doesn't exist yet and will come in due time, but it will not affect the overall outcome.

Once the integration is complete we will relook any inconsistencies and fix up any additional functionality required.

Signed-off-by: Kenneth Lim <kenlimmj@gmail.com>